fix device consume

+++ src/server/service/trade.go
@@ -50,7 +50,7 @@ func (self *TradeService) BasicOfDevice(serialNumber string, mobile string, page
 			m["address"] = device.Address
 		}
 		m["token"] = ticket.Token
-		m["amount"] = ticket.Value / 100
+		m["amount"] = float64(ticket.Value) / 100
 		m["account"] = ticket.Mobile
 		m["pulseType"] = 600 + ticket.DeviceMode
 		m["createdAt"] = ticket.CreatedAt
